Move is_two_way() into a trait, since it makes more sense there
