+++ src/transformation/transform_base64.rs
@@ -155,6 +155,12 @@ impl TransformerTrait for TransformBase64 {
             true  => self.check_permissive(buffer),
         }
     }
+
+    fn is_two_way(&self) -> bool {
+        // It's two-way if it's not permissive (permissive allows illegal stuff,
+        // which would only be one way).
+        !self.settings.permissive
+    }
 }
 
 #[cfg(test)]
