+++ src/transformation/transform_hex.rs
@@ -29,6 +29,10 @@ impl TransformerTrait for TransformHex {
     fn untransform(&self, buffer: &Vec<u8>) -> SimpleResult<Vec<u8>> {
         Ok(hex::encode(buffer).into_bytes())
     }
+
+    fn is_two_way(&self) -> bool {
+        true
+    }
 }
 
 #[cfg(test)]
